requirements.txt
asgiref==3.10.0
Django==5.0
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==25.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
